Split store into lib, local, remote modules
These are currently unreasonably big in one file.
