+++ src/rust/engine/fs/store/src/lib.rs
@@ -58,6 +58,9 @@ const GIGABYTES: usize = 1024 * MEGABYTES;
 // after garbage collection. We almost certainly want to make this configurable.
 pub const DEFAULT_LOCAL_STORE_GC_TARGET_BYTES: usize = 4 * GIGABYTES;
 
+mod local;
+mod remote;
+
 // Summary of the files and directories uploaded with an operation
 // ingested_file_{count, bytes}: Number and combined size of processed files
 // uploaded_file_{count, bytes}: Number and combined size of files uploaded to the remote
@@ -945,1601 +948,6 @@ impl Store {
 pub enum EntryType {
   Directory,
   File,
-}
-
-mod local {
-  use super::{EntryType, ShrinkBehavior, GIGABYTES};
-
-  use boxfuture::{try_future, BoxFuture, Boxable};
-  use bytes::Bytes;
-  use digest::{Digest as DigestTrait, FixedOutput};
-  use futures::future::{self, Future};
-  use hashing::{Digest, Fingerprint, EMPTY_DIGEST};
-  use lmdb::Error::NotFound;
-  use lmdb::{self, Cursor, Database, RwTransaction, Transaction, WriteFlags};
-  use sha2::Sha256;
-  use sharded_lmdb::ShardedLmdb;
-  use std;
-  use std::collections::BinaryHeap;
-  use std::path::Path;
-  use std::sync::Arc;
-  use std::time;
-
-  #[derive(Clone)]
-  pub struct ByteStore {
-    inner: Arc<InnerStore>,
-  }
-
-  struct InnerStore {
-    // Store directories separately from files because:
-    //  1. They may have different lifetimes.
-    //  2. It's nice to know whether we should be able to parse something as a proto.
-    file_dbs: Result<Arc<ShardedLmdb>, String>,
-    directory_dbs: Result<Arc<ShardedLmdb>, String>,
-    executor: task_executor::Executor,
-  }
-
-  impl ByteStore {
-    pub fn new<P: AsRef<Path>>(
-      executor: task_executor::Executor,
-      path: P,
-    ) -> Result<ByteStore, String> {
-      let root = path.as_ref();
-      let files_root = root.join("files");
-      let directories_root = root.join("directories");
-      Ok(ByteStore {
-        inner: Arc::new(InnerStore {
-          // We want these stores to be allowed to grow very large, in case we are on a system with
-          // large disks which doesn't want to GC a lot.
-          // It doesn't reflect space allocated on disk, or RAM allocated (it may be reflected in
-          // VIRT but not RSS). There is no practical upper bound on this number, so we set them
-          // ridiculously high.
-          // However! We set them lower than we'd otherwise choose because sometimes we see tests on
-          // travis fail because they can't allocate virtual memory, if there are multiple Stores
-          // in memory at the same time. We don't know why they're not efficiently garbage collected
-          // by python, but they're not, so...
-          file_dbs: ShardedLmdb::new(files_root.clone(), 100 * GIGABYTES, executor.clone())
-            .map(Arc::new),
-          directory_dbs: ShardedLmdb::new(
-            directories_root.clone(),
-            5 * GIGABYTES,
-            executor.clone(),
-          )
-          .map(Arc::new),
-          executor: executor,
-        }),
-      })
-    }
-
-    // Note: This performs IO on the calling thread. Hopefully the IO is small enough not to matter.
-    pub fn entry_type(&self, fingerprint: &Fingerprint) -> Result<Option<EntryType>, String> {
-      if *fingerprint == EMPTY_DIGEST.0 {
-        // Technically this is valid as both; choose Directory in case a caller is checking whether
-        // it _can_ be a Directory.
-        return Ok(Some(EntryType::Directory));
-      }
-      {
-        let (env, directory_database, _) = self.inner.directory_dbs.clone()?.get(fingerprint);
-        let txn = env
-          .begin_ro_txn()
-          .map_err(|err| format!("Failed to begin read transaction: {:?}", err))?;
-        match txn.get(directory_database, &fingerprint.as_ref()) {
-          Ok(_) => return Ok(Some(EntryType::Directory)),
-          Err(NotFound) => {}
-          Err(err) => {
-            return Err(format!(
-              "Error reading from store when determining type of fingerprint {}: {}",
-              fingerprint, err
-            ));
-          }
-        };
-      }
-      let (env, file_database, _) = self.inner.file_dbs.clone()?.get(fingerprint);
-      let txn = env
-        .begin_ro_txn()
-        .map_err(|err| format!("Failed to begin read transaction: {}", err))?;
-      match txn.get(file_database, &fingerprint.as_ref()) {
-        Ok(_) => return Ok(Some(EntryType::File)),
-        Err(NotFound) => {}
-        Err(err) => {
-          return Err(format!(
-            "Error reading from store when determining type of fingerprint {}: {}",
-            fingerprint, err
-          ));
-        }
-      };
-      Ok(None)
-    }
-
-    pub fn lease_all<'a, Ds: Iterator<Item = &'a Digest>>(
-      &self,
-      digests: Ds,
-    ) -> Result<(), String> {
-      let until = Self::default_lease_until_secs_since_epoch();
-      for digest in digests {
-        let (env, _, lease_database) = self.inner.file_dbs.clone()?.get(&digest.0);
-        env
-          .begin_rw_txn()
-          .and_then(|mut txn| self.lease(lease_database, &digest.0, until, &mut txn))
-          .map_err(|err| format!("Error leasing digest {:?}: {}", digest, err))?;
-      }
-      Ok(())
-    }
-
-    fn default_lease_until_secs_since_epoch() -> u64 {
-      let now_since_epoch = time::SystemTime::now()
-        .duration_since(time::UNIX_EPOCH)
-        .expect("Surely you're not before the unix epoch?");
-      (now_since_epoch + time::Duration::from_secs(2 * 60 * 60)).as_secs()
-    }
-
-    fn lease(
-      &self,
-      database: Database,
-      fingerprint: &Fingerprint,
-      until_secs_since_epoch: u64,
-      txn: &mut RwTransaction<'_>,
-    ) -> Result<(), lmdb::Error> {
-      txn.put(
-        database,
-        &fingerprint.as_ref(),
-        &until_secs_since_epoch.to_le_bytes(),
-        WriteFlags::empty(),
-      )
-    }
-
-    ///
-    /// Attempts to shrink the stored files to be no bigger than target_bytes
-    /// (excluding lmdb overhead).
-    ///
-    /// Returns the size it was shrunk to, which may be larger than target_bytes.
-    ///
-    /// Ignores directories. TODO: Shrink directories.
-    ///
-    /// TODO: Use LMDB database statistics when lmdb-rs exposes them.
-    ///
-    pub fn shrink(
-      &self,
-      target_bytes: usize,
-      shrink_behavior: ShrinkBehavior,
-    ) -> Result<usize, String> {
-      let mut used_bytes: usize = 0;
-      let mut fingerprints_by_expired_ago = BinaryHeap::new();
-
-      self.aged_fingerprints(
-        EntryType::File,
-        &mut used_bytes,
-        &mut fingerprints_by_expired_ago,
-      )?;
-      self.aged_fingerprints(
-        EntryType::Directory,
-        &mut used_bytes,
-        &mut fingerprints_by_expired_ago,
-      )?;
-      while used_bytes > target_bytes {
-        let aged_fingerprint = fingerprints_by_expired_ago
-          .pop()
-          .expect("lmdb corruption detected, sum of size of blobs exceeded stored blobs");
-        if aged_fingerprint.expired_seconds_ago == 0 {
-          // Ran out of expired blobs - everything remaining is leased and cannot be collected.
-          return Ok(used_bytes);
-        }
-        let lmdbs = match aged_fingerprint.entry_type {
-          EntryType::File => self.inner.file_dbs.clone(),
-          EntryType::Directory => self.inner.directory_dbs.clone(),
-        };
-        let (env, database, lease_database) = lmdbs.clone()?.get(&aged_fingerprint.fingerprint);
-        {
-          env
-            .begin_rw_txn()
-            .and_then(|mut txn| {
-              txn.del(database, &aged_fingerprint.fingerprint.as_ref(), None)?;
-
-              txn
-                .del(lease_database, &aged_fingerprint.fingerprint.as_ref(), None)
-                .or_else(|err| match err {
-                  NotFound => Ok(()),
-                  err => Err(err),
-                })?;
-              used_bytes -= aged_fingerprint.size_bytes;
-              txn.commit()
-            })
-            .map_err(|err| format!("Error garbage collecting: {}", err))?;
-        }
-      }
-
-      if shrink_behavior == ShrinkBehavior::Compact {
-        self.inner.file_dbs.clone()?.compact()?;
-      }
-
-      Ok(used_bytes)
-    }
-
-    fn aged_fingerprints(
-      &self,
-      entry_type: EntryType,
-      used_bytes: &mut usize,
-      fingerprints_by_expired_ago: &mut BinaryHeap<AgedFingerprint>,
-    ) -> Result<(), String> {
-      let database = match entry_type {
-        EntryType::File => self.inner.file_dbs.clone(),
-        EntryType::Directory => self.inner.directory_dbs.clone(),
-      };
-
-      for &(ref env, ref database, ref lease_database) in &database?.all_lmdbs() {
-        let txn = env
-          .begin_ro_txn()
-          .map_err(|err| format!("Error beginning transaction to garbage collect: {}", err))?;
-        let mut cursor = txn
-          .open_ro_cursor(*database)
-          .map_err(|err| format!("Failed to open lmdb read cursor: {}", err))?;
-        for (key, bytes) in cursor.iter() {
-          *used_bytes += bytes.len();
-
-          // Random access into the lease_database is slower than iterating, but hopefully garbage
-          // collection is rare enough that we can get away with this, rather than do two passes
-          // here (either to populate leases into pre-populated AgedFingerprints, or to read sizes
-          // when we delete from lmdb to track how much we've freed).
-          let lease_until_unix_timestamp = txn
-            .get(*lease_database, &key)
-            .map(|b| {
-              let mut array = [0_u8; 8];
-              array.copy_from_slice(b);
-              u64::from_le_bytes(array)
-            })
-            .unwrap_or_else(|e| match e {
-              NotFound => 0,
-              e => panic!("Error reading lease, probable lmdb corruption: {:?}", e),
-            });
-
-          let leased_until =
-            time::UNIX_EPOCH + time::Duration::from_secs(lease_until_unix_timestamp);
-
-          let expired_seconds_ago = time::SystemTime::now()
-            .duration_since(leased_until)
-            .map(|t| t.as_secs())
-            // 0 indicates unleased.
-            .unwrap_or(0);
-
-          fingerprints_by_expired_ago.push(AgedFingerprint {
-            expired_seconds_ago: expired_seconds_ago,
-            fingerprint: Fingerprint::from_bytes_unsafe(key),
-            size_bytes: bytes.len(),
-            entry_type: entry_type,
-          });
-        }
-      }
-      Ok(())
-    }
-
-    pub fn store_bytes(
-      &self,
-      entry_type: EntryType,
-      bytes: Bytes,
-      initial_lease: bool,
-    ) -> impl Future<Item = Digest, Error = String> {
-      let dbs = match entry_type {
-        EntryType::Directory => self.inner.directory_dbs.clone(),
-        EntryType::File => self.inner.file_dbs.clone(),
-      };
-      let bytes2 = bytes.clone();
-      self
-        .inner
-        .executor
-        .spawn_on_io_pool(futures::future::lazy(move || {
-          let fingerprint = {
-            let mut hasher = Sha256::default();
-            hasher.input(&bytes);
-            Fingerprint::from_bytes_unsafe(hasher.fixed_result().as_slice())
-          };
-          Ok(Digest(fingerprint, bytes.len()))
-        }))
-        .and_then(move |digest| {
-          future::done(dbs)
-            .and_then(move |db| db.store_bytes(digest.0, bytes2, initial_lease))
-            .map(move |()| digest)
-        })
-    }
-
-    pub fn load_bytes_with<T: Send + 'static, F: Fn(Bytes) -> T + Send + Sync + 'static>(
-      &self,
-      entry_type: EntryType,
-      digest: Digest,
-      f: F,
-    ) -> BoxFuture<Option<T>, String> {
-      if digest == EMPTY_DIGEST {
-        // Avoid expensive I/O for this super common case.
-        // Also, this allows some client-provided operations (like merging snapshots) to work
-        // without needing to first store the empty snapshot.
-        return future::ok(Some(f(Bytes::new()))).to_boxed();
-      }
-
-      let dbs = match entry_type {
-        EntryType::Directory => self.inner.directory_dbs.clone(),
-        EntryType::File => self.inner.file_dbs.clone(),
-      };
-
-      try_future!(dbs).load_bytes_with(digest.0, move |bytes| {
-        if bytes.len() == digest.1 {
-          Ok(f(bytes))
-        } else {
-          Err(format!("Got hash collision reading from store - digest {:?} was requested, but retrieved bytes with that fingerprint had length {}. Congratulations, you may have broken sha256! Underlying bytes: {:?}", digest, bytes.len(), bytes))
-        }
-      }).to_boxed()
-    }
-
-    pub fn all_digests(&self, entry_type: EntryType) -> Result<Vec<Digest>, String> {
-      let database = match entry_type {
-        EntryType::File => self.inner.file_dbs.clone(),
-        EntryType::Directory => self.inner.directory_dbs.clone(),
-      };
-      let mut digests = vec![];
-      for &(ref env, ref database, ref _lease_database) in &database?.all_lmdbs() {
-        let txn = env
-          .begin_ro_txn()
-          .map_err(|err| format!("Error beginning transaction to garbage collect: {}", err))?;
-        let mut cursor = txn
-          .open_ro_cursor(*database)
-          .map_err(|err| format!("Failed to open lmdb read cursor: {}", err))?;
-        for (key, bytes) in cursor.iter() {
-          digests.push(Digest(Fingerprint::from_bytes_unsafe(key), bytes.len()));
-        }
-      }
-      Ok(digests)
-    }
-  }
-
-  #[derive(Eq, PartialEq, Ord, PartialOrd)]
-  struct AgedFingerprint {
-    // expired_seconds_ago must be the first field for the Ord implementation.
-    expired_seconds_ago: u64,
-    fingerprint: Fingerprint,
-    size_bytes: usize,
-    entry_type: EntryType,
-  }
-
-  #[cfg(test)]
-  pub mod tests {
-    use super::super::tests::block_on;
-    use super::{ByteStore, EntryType, ShrinkBehavior};
-    use bytes::{BufMut, Bytes, BytesMut};
-    use hashing::{Digest, Fingerprint};
-    use std::path::Path;
-    use tempfile::TempDir;
-    use testutil::data::{TestData, TestDirectory};
-    use walkdir::WalkDir;
-
-    #[test]
-    fn save_file() {
-      let dir = TempDir::new().unwrap();
-
-      let testdata = TestData::roland();
-      assert_eq!(
-        block_on(new_store(dir.path()).store_bytes(EntryType::File, testdata.bytes(), false,)),
-        Ok(testdata.digest())
-      );
-    }
-
-    #[test]
-    fn save_file_is_idempotent() {
-      let dir = TempDir::new().unwrap();
-
-      let testdata = TestData::roland();
-      block_on(new_store(dir.path()).store_bytes(EntryType::File, testdata.bytes(), false))
-        .unwrap();
-      assert_eq!(
-        block_on(new_store(dir.path()).store_bytes(EntryType::File, testdata.bytes(), false,)),
-        Ok(testdata.digest())
-      );
-    }
-
-    #[test]
-    fn roundtrip_file() {
-      let testdata = TestData::roland();
-      let dir = TempDir::new().unwrap();
-
-      let store = new_store(dir.path());
-      let hash = prime_store_with_file_bytes(&store, testdata.bytes());
-      assert_eq!(load_file_bytes(&store, hash), Ok(Some(testdata.bytes())));
-    }
-
-    #[test]
-    fn missing_file() {
-      let dir = TempDir::new().unwrap();
-      assert_eq!(
-        load_file_bytes(&new_store(dir.path()), TestData::roland().digest()),
-        Ok(None)
-      );
-    }
-
-    #[test]
-    fn record_and_load_directory_proto() {
-      let dir = TempDir::new().unwrap();
-      let testdir = TestDirectory::containing_roland();
-
-      assert_eq!(
-        block_on(new_store(dir.path()).store_bytes(EntryType::Directory, testdir.bytes(), false,)),
-        Ok(testdir.digest())
-      );
-
-      assert_eq!(
-        load_directory_proto_bytes(&new_store(dir.path()), testdir.digest()),
-        Ok(Some(testdir.bytes()))
-      );
-    }
-
-    #[test]
-    fn missing_directory() {
-      let dir = TempDir::new().unwrap();
-      let testdir = TestDirectory::containing_roland();
-
-      assert_eq!(
-        load_directory_proto_bytes(&new_store(dir.path()), testdir.digest()),
-        Ok(None)
-      );
-    }
-
-    #[test]
-    fn file_is_not_directory_proto() {
-      let dir = TempDir::new().unwrap();
-      let testdata = TestData::roland();
-
-      block_on(new_store(dir.path()).store_bytes(EntryType::File, testdata.bytes(), false))
-        .unwrap();
-
-      assert_eq!(
-        load_directory_proto_bytes(&new_store(dir.path()), testdata.digest()),
-        Ok(None)
-      );
-    }
-
-    #[test]
-    fn garbage_collect_nothing_to_do() {
-      let dir = TempDir::new().unwrap();
-      let store = new_store(dir.path());
-      let bytes = Bytes::from("0123456789");
-      block_on(store.store_bytes(EntryType::File, bytes.clone(), false)).expect("Error storing");
-      store
-        .shrink(10, ShrinkBehavior::Fast)
-        .expect("Error shrinking");
-      assert_eq!(
-        load_bytes(
-          &store,
-          EntryType::File,
-          Digest(
-            Fingerprint::from_hex_string(
-              "84d89877f0d4041efb6bf91a16f0248f2fd573e6af05c19f96bedb9f882f7882",
-            )
-            .unwrap(),
-            10
-          )
-        ),
-        Ok(Some(bytes))
-      );
-    }
-
-    #[test]
-    fn garbage_collect_nothing_to_do_with_lease() {
-      let dir = TempDir::new().unwrap();
-      let store = new_store(dir.path());
-      let bytes = Bytes::from("0123456789");
-      block_on(store.store_bytes(EntryType::File, bytes.clone(), false)).expect("Error storing");
-      let file_fingerprint = Fingerprint::from_hex_string(
-        "84d89877f0d4041efb6bf91a16f0248f2fd573e6af05c19f96bedb9f882f7882",
-      )
-      .unwrap();
-      let file_digest = Digest(file_fingerprint, 10);
-      store
-        .lease_all(vec![file_digest].iter())
-        .expect("Error leasing");
-      store
-        .shrink(10, ShrinkBehavior::Fast)
-        .expect("Error shrinking");
-      assert_eq!(
-        load_bytes(&store, EntryType::File, file_digest),
-        Ok(Some(bytes))
-      );
-    }
-
-    #[test]
-    fn garbage_collect_remove_one_of_two_files_no_leases() {
-      let dir = TempDir::new().unwrap();
-      let store = new_store(dir.path());
-      let bytes_1 = Bytes::from("0123456789");
-      let fingerprint_1 = Fingerprint::from_hex_string(
-        "84d89877f0d4041efb6bf91a16f0248f2fd573e6af05c19f96bedb9f882f7882",
-      )
-      .unwrap();
-      let digest_1 = Digest(fingerprint_1, 10);
-      let bytes_2 = Bytes::from("9876543210");
-      let fingerprint_2 = Fingerprint::from_hex_string(
-        "7619ee8cea49187f309616e30ecf54be072259b43760f1f550a644945d5572f2",
-      )
-      .unwrap();
-      let digest_2 = Digest(fingerprint_2, 10);
-      block_on(store.store_bytes(EntryType::File, bytes_1.clone(), false)).expect("Error storing");
-      block_on(store.store_bytes(EntryType::File, bytes_2.clone(), false)).expect("Error storing");
-      store
-        .shrink(10, ShrinkBehavior::Fast)
-        .expect("Error shrinking");
-      let mut entries = Vec::new();
-      entries.push(load_bytes(&store, EntryType::File, digest_1).expect("Error loading bytes"));
-      entries.push(load_bytes(&store, EntryType::File, digest_2).expect("Error loading bytes"));
-      assert_eq!(
-        1,
-        entries.iter().filter(|maybe| maybe.is_some()).count(),
-        "Want one Some but got: {:?}",
-        entries
-      );
-    }
-
-    #[test]
-    fn garbage_collect_remove_both_files_no_leases() {
-      let dir = TempDir::new().unwrap();
-      let store = new_store(dir.path());
-      let bytes_1 = Bytes::from("0123456789");
-      let fingerprint_1 = Fingerprint::from_hex_string(
-        "84d89877f0d4041efb6bf91a16f0248f2fd573e6af05c19f96bedb9f882f7882",
-      )
-      .unwrap();
-      let digest_1 = Digest(fingerprint_1, 10);
-      let bytes_2 = Bytes::from("9876543210");
-      let fingerprint_2 = Fingerprint::from_hex_string(
-        "7619ee8cea49187f309616e30ecf54be072259b43760f1f550a644945d5572f2",
-      )
-      .unwrap();
-      let digest_2 = Digest(fingerprint_2, 10);
-      block_on(store.store_bytes(EntryType::File, bytes_1.clone(), false)).expect("Error storing");
-      block_on(store.store_bytes(EntryType::File, bytes_2.clone(), false)).expect("Error storing");
-      store
-        .shrink(1, ShrinkBehavior::Fast)
-        .expect("Error shrinking");
-      assert_eq!(
-        load_bytes(&store, EntryType::File, digest_1),
-        Ok(None),
-        "Should have garbage collected {:?}",
-        fingerprint_1
-      );
-      assert_eq!(
-        load_bytes(&store, EntryType::File, digest_2),
-        Ok(None),
-        "Should have garbage collected {:?}",
-        fingerprint_2
-      );
-    }
-
-    #[test]
-    fn garbage_collect_remove_one_of_two_directories_no_leases() {
-      let dir = TempDir::new().unwrap();
-
-      let testdir = TestDirectory::containing_roland();
-      let other_testdir = TestDirectory::containing_dnalor();
-
-      let store = new_store(dir.path());
-      block_on(store.store_bytes(EntryType::Directory, testdir.bytes(), false))
-        .expect("Error storing");
-      block_on(store.store_bytes(EntryType::Directory, other_testdir.bytes(), false))
-        .expect("Error storing");
-      store
-        .shrink(80, ShrinkBehavior::Fast)
-        .expect("Error shrinking");
-      let mut entries = Vec::new();
-      entries.push(
-        load_bytes(&store, EntryType::Directory, testdir.digest()).expect("Error loading bytes"),
-      );
-      entries.push(
-        load_bytes(&store, EntryType::Directory, other_testdir.digest())
-          .expect("Error loading bytes"),
-      );
-      assert_eq!(
-        1,
-        entries.iter().filter(|maybe| maybe.is_some()).count(),
-        "Want one Some but got: {:?}",
-        entries
-      );
-    }
-
-    #[test]
-    fn garbage_collect_remove_file_with_leased_directory() {
-      let dir = TempDir::new().unwrap();
-      let store = new_store(dir.path());
-
-      let testdir = TestDirectory::containing_roland();
-      let testdata = TestData::fourty_chars();
-
-      block_on(store.store_bytes(EntryType::Directory, testdir.bytes(), true))
-        .expect("Error storing");
-
-      block_on(store.store_bytes(EntryType::File, testdata.bytes(), false)).expect("Error storing");
-
-      store
-        .shrink(80, ShrinkBehavior::Fast)
-        .expect("Error shrinking");
-
-      assert_eq!(
-        load_bytes(&store, EntryType::File, testdata.digest()),
-        Ok(None),
-        "File was present when it should've been garbage collected"
-      );
-      assert_eq!(
-        load_bytes(&store, EntryType::Directory, testdir.digest()),
-        Ok(Some(testdir.bytes())),
-        "Directory was missing despite lease"
-      );
-    }
-
-    #[test]
-    fn garbage_collect_remove_file_while_leased_file() {
-      let dir = TempDir::new().unwrap();
-      let store = new_store(dir.path());
-
-      let testdir = TestDirectory::containing_roland();
-
-      block_on(store.store_bytes(EntryType::Directory, testdir.bytes(), false))
-        .expect("Error storing");
-      let fourty_chars = TestData::fourty_chars();
-      block_on(store.store_bytes(EntryType::File, fourty_chars.bytes(), true))
-        .expect("Error storing");
-
-      store
-        .shrink(80, ShrinkBehavior::Fast)
-        .expect("Error shrinking");
-
-      assert_eq!(
-        load_bytes(&store, EntryType::File, fourty_chars.digest()),
-        Ok(Some(fourty_chars.bytes())),
-        "File was missing despite lease"
-      );
-      assert_eq!(
-        load_bytes(&store, EntryType::Directory, testdir.digest()),
-        Ok(None),
-        "Directory was present when it should've been garbage collected"
-      );
-    }
-
-    #[test]
-    fn garbage_collect_fail_because_too_many_leases() {
-      let dir = TempDir::new().unwrap();
-      let store = new_store(dir.path());
-
-      let testdir = TestDirectory::containing_roland();
-      let fourty_chars = TestData::fourty_chars();
-
-      block_on(store.store_bytes(EntryType::Directory, testdir.bytes(), true))
-        .expect("Error storing");
-      block_on(store.store_bytes(EntryType::File, fourty_chars.bytes(), true))
-        .expect("Error storing");
-
-      block_on(store.store_bytes(EntryType::File, TestData::roland().bytes(), false))
-        .expect("Error storing");
-
-      assert_eq!(store.shrink(80, ShrinkBehavior::Fast), Ok(160));
-
-      assert_eq!(
-        load_bytes(&store, EntryType::File, fourty_chars.digest()),
-        Ok(Some(fourty_chars.bytes())),
-        "Leased file should still be present"
-      );
-      assert_eq!(
-        load_bytes(&store, EntryType::Directory, testdir.digest()),
-        Ok(Some(testdir.bytes())),
-        "Leased directory should still be present"
-      );
-      // Whether the unleased file is present is undefined.
-    }
-
-    #[test]
-    fn garbage_collect_and_compact() {
-      let dir = TempDir::new().unwrap();
-      let store = new_store(dir.path());
-
-      let write_one_meg = |byte: u8| {
-        let mut bytes = BytesMut::with_capacity(1024 * 1024);
-        for _ in 0..1024 * 1024 {
-          bytes.put(byte);
-        }
-        block_on(store.store_bytes(EntryType::File, bytes.freeze(), false)).expect("Error storing");
-      };
-
-      write_one_meg(b'0');
-
-      write_one_meg(b'1');
-
-      let size = get_directory_size(dir.path());
-      assert!(
-        size >= 2 * 1024 * 1024,
-        "Expect size to be at least 2MB but was {}",
-        size
-      );
-
-      store
-        .shrink(1024 * 1024, ShrinkBehavior::Compact)
-        .expect("Error shrinking");
-
-      let size = get_directory_size(dir.path());
-      assert!(
-        size < 2 * 1024 * 1024,
-        "Expect size to be less than 2MB but was {}",
-        size
-      );
-    }
-
-    #[test]
-    fn entry_type_for_file() {
-      let testdata = TestData::roland();
-      let testdir = TestDirectory::containing_roland();
-      let dir = TempDir::new().unwrap();
-      let store = new_store(dir.path());
-      block_on(store.store_bytes(EntryType::Directory, testdir.bytes(), false))
-        .expect("Error storing");
-      prime_store_with_file_bytes(&store, testdata.bytes());
-      assert_eq!(
-        store.entry_type(&testdata.fingerprint()),
-        Ok(Some(EntryType::File))
-      )
-    }
-
-    #[test]
-    fn entry_type_for_directory() {
-      let testdata = TestData::roland();
-      let testdir = TestDirectory::containing_roland();
-      let dir = TempDir::new().unwrap();
-      let store = new_store(dir.path());
-      block_on(store.store_bytes(EntryType::Directory, testdir.bytes(), false))
-        .expect("Error storing");
-      prime_store_with_file_bytes(&store, testdata.bytes());
-      assert_eq!(
-        store.entry_type(&testdir.fingerprint()),
-        Ok(Some(EntryType::Directory))
-      )
-    }
-
-    #[test]
-    fn entry_type_for_missing() {
-      let testdata = TestData::roland();
-      let testdir = TestDirectory::containing_roland();
-      let dir = TempDir::new().unwrap();
-      let store = new_store(dir.path());
-      block_on(store.store_bytes(EntryType::Directory, testdir.bytes(), false))
-        .expect("Error storing");
-      prime_store_with_file_bytes(&store, testdata.bytes());
-      assert_eq!(
-        store.entry_type(&TestDirectory::recursive().fingerprint()),
-        Ok(None)
-      )
-    }
-
-    #[test]
-    pub fn empty_file_is_known() {
-      let dir = TempDir::new().unwrap();
-      let store = new_store(dir.path());
-      let empty_file = TestData::empty();
-      assert_eq!(
-        block_on(store.load_bytes_with(EntryType::File, empty_file.digest(), |b| b)),
-        Ok(Some(empty_file.bytes())),
-      )
-    }
-
-    #[test]
-    pub fn empty_directory_is_known() {
-      let dir = TempDir::new().unwrap();
-      let store = new_store(dir.path());
-      let empty_dir = TestDirectory::empty();
-      assert_eq!(
-        block_on(store.load_bytes_with(EntryType::Directory, empty_dir.digest(), |b| b)),
-        Ok(Some(empty_dir.bytes())),
-      )
-    }
-
-    #[test]
-    pub fn all_digests() {
-      let dir = TempDir::new().unwrap();
-      let store = new_store(dir.path());
-      let digest = prime_store_with_file_bytes(&store, TestData::roland().bytes());
-      assert_eq!(Ok(vec![digest]), store.all_digests(EntryType::File));
-    }
-
-    pub fn new_store<P: AsRef<Path>>(dir: P) -> ByteStore {
-      ByteStore::new(task_executor::Executor::new(), dir).unwrap()
-    }
-
-    pub fn load_file_bytes(store: &ByteStore, digest: Digest) -> Result<Option<Bytes>, String> {
-      load_bytes(&store, EntryType::File, digest)
-    }
-
-    pub fn load_directory_proto_bytes(
-      store: &ByteStore,
-      digest: Digest,
-    ) -> Result<Option<Bytes>, String> {
-      load_bytes(&store, EntryType::Directory, digest)
-    }
-
-    fn load_bytes(
-      store: &ByteStore,
-      entry_type: EntryType,
-      digest: Digest,
-    ) -> Result<Option<Bytes>, String> {
-      block_on(store.load_bytes_with(entry_type, digest, |b| b))
-    }
-
-    fn prime_store_with_file_bytes(store: &ByteStore, bytes: Bytes) -> Digest {
-      block_on(store.store_bytes(EntryType::File, bytes, false)).expect("Error storing file bytes")
-    }
-
-    fn get_directory_size(path: &Path) -> usize {
-      let mut len: usize = 0;
-      for entry in WalkDir::new(path) {
-        len += entry
-          .expect("Error walking directory")
-          .metadata()
-          .expect("Error reading metadata")
-          .len() as usize;
-      }
-      len
-    }
-  }
-}
-
-mod remote {
-  use super::{BackoffConfig, EntryType};
-
-  use bazel_protos;
-  use boxfuture::{BoxFuture, Boxable};
-  use bytes::{Bytes, BytesMut};
-  use concrete_time::TimeSpan;
-  use digest::{Digest as DigestTrait, FixedOutput};
-  use futures::{self, future, Future, IntoFuture, Sink, Stream};
-  use grpcio;
-  use hashing::{Digest, Fingerprint};
-  use serverset::{Retry, Serverset};
-  use sha2::Sha256;
-  use std::cmp::min;
-  use std::collections::HashSet;
-  use std::sync::Arc;
-  use std::time::Duration;
-  use uuid;
-  use workunit_store::WorkUnitStore;
-
-  #[derive(Clone)]
-  pub struct ByteStore {
-    instance_name: Option<String>,
-    chunk_size_bytes: usize,
-    upload_timeout: Duration,
-    rpc_attempts: usize,
-    env: Arc<grpcio::Environment>,
-    serverset: Serverset<grpcio::Channel>,
-    authorization_header: Option<String>,
-  }
-
-  impl ByteStore {
-    pub fn new(
-      cas_addresses: Vec<String>,
-      instance_name: Option<String>,
-      root_ca_certs: Option<Vec<u8>>,
-      oauth_bearer_token: Option<String>,
-      thread_count: usize,
-      chunk_size_bytes: usize,
-      upload_timeout: Duration,
-      backoff_config: BackoffConfig,
-      rpc_retries: usize,
-      connection_limit: usize,
-    ) -> Result<ByteStore, String> {
-      let env = Arc::new(grpcio::Environment::new(thread_count));
-      let env2 = env.clone();
-
-      let connect = move |cas_address: &str| {
-        let builder = grpcio::ChannelBuilder::new(env2.clone());
-        if let Some(ref root_ca_certs) = root_ca_certs {
-          let creds = grpcio::ChannelCredentialsBuilder::new()
-            .root_cert(root_ca_certs.clone())
-            .build();
-          builder.secure_connect(cas_address, creds)
-        } else {
-          builder.connect(cas_address)
-        }
-      };
-
-      let serverset = Serverset::new(cas_addresses, connect, connection_limit, backoff_config)?;
-
-      Ok(ByteStore {
-        instance_name,
-        chunk_size_bytes,
-        upload_timeout,
-        rpc_attempts: rpc_retries + 1,
-        env,
-        serverset,
-        authorization_header: oauth_bearer_token.map(|t| format!("Bearer {}", t)),
-      })
-    }
-
-    fn with_byte_stream_client<
-      Value: Send + 'static,
-      Fut: Future<Item = Value, Error = String>,
-      IntoFut: IntoFuture<Future = Fut, Item = Value, Error = String>,
-      F: Fn(bazel_protos::bytestream_grpc::ByteStreamClient) -> IntoFut
-        + Send
-        + Sync
-        + Clone
-        + 'static,
-    >(
-      &self,
-      f: F,
-    ) -> impl Future<Item = Value, Error = String> {
-      Retry(self.serverset.clone()).all_errors_immediately(
-        move |channel| {
-          f(bazel_protos::bytestream_grpc::ByteStreamClient::new(
-            channel,
-          ))
-        },
-        self.rpc_attempts,
-      )
-    }
-
-    fn with_cas_client<
-      Value: Send + 'static,
-      Fut: Future<Item = Value, Error = String>,
-      IntoFut: IntoFuture<Future = Fut, Item = Value, Error = String>,
-      F: Fn(bazel_protos::remote_execution_grpc::ContentAddressableStorageClient) -> IntoFut
-        + Send
-        + Sync
-        + Clone
-        + 'static,
-    >(
-      &self,
-      f: F,
-    ) -> impl Future<Item = Value, Error = String> {
-      Retry(self.serverset.clone()).all_errors_immediately(
-        move |channel| {
-          f(bazel_protos::remote_execution_grpc::ContentAddressableStorageClient::new(channel))
-        },
-        self.rpc_attempts,
-      )
-    }
-
-    fn call_option(&self) -> grpcio::CallOption {
-      let mut call_option = grpcio::CallOption::default();
-      if let Some(ref authorization_header) = self.authorization_header {
-        let mut builder = grpcio::MetadataBuilder::with_capacity(1);
-        builder
-          .add_str("authorization", &authorization_header)
-          .unwrap();
-        call_option = call_option.headers(builder.build());
-      }
-      call_option
-    }
-
-    pub fn store_bytes(
-      &self,
-      bytes: Bytes,
-      workunit_store: WorkUnitStore,
-    ) -> BoxFuture<Digest, String> {
-      let start_time = std::time::SystemTime::now();
-
-      let mut hasher = Sha256::default();
-      hasher.input(&bytes);
-      let fingerprint = Fingerprint::from_bytes_unsafe(hasher.fixed_result().as_slice());
-      let len = bytes.len();
-      let digest = Digest(fingerprint, len);
-      let resource_name = format!(
-        "{}/uploads/{}/blobs/{}/{}",
-        self.instance_name.clone().unwrap_or_default(),
-        uuid::Uuid::new_v4(),
-        digest.0,
-        digest.1,
-      );
-      let workunit_name = format!("store_bytes({})", resource_name.clone());
-      let workunit_store = workunit_store.clone();
-      let store = self.clone();
-      self
-        .with_byte_stream_client(move |client| {
-          match client
-            .write_opt(store.call_option().timeout(store.upload_timeout))
-            .map(|v| (v, client))
-          {
-            Err(err) => future::err(format!(
-              "Error attempting to connect to upload digest {:?}: {:?}",
-              digest, err
-            ))
-            .to_boxed(),
-            Ok(((sender, receiver), _client)) => {
-              let chunk_size_bytes = store.chunk_size_bytes;
-              let resource_name = resource_name.clone();
-              let bytes = bytes.clone();
-              let stream = futures::stream::unfold::<
-                _,
-                _,
-                futures::future::FutureResult<_, grpcio::Error>,
-                _,
-              >((0, false), move |(offset, has_sent_any)| {
-                if offset >= bytes.len() && has_sent_any {
-                  None
-                } else {
-                  let mut req = bazel_protos::bytestream::WriteRequest::new();
-                  req.set_resource_name(resource_name.clone());
-                  req.set_write_offset(offset as i64);
-                  let next_offset = min(offset + chunk_size_bytes, bytes.len());
-                  req.set_finish_write(next_offset == bytes.len());
-                  req.set_data(bytes.slice(offset, next_offset));
-                  Some(future::ok((
-                    (req, grpcio::WriteFlags::default()),
-                    (next_offset, true),
-                  )))
-                }
-              });
-
-              sender
-                .send_all(stream)
-                .map(|_| ())
-                .or_else(move |e| {
-                  match e {
-                    // Some implementations of the remote execution API early-return if the blob has
-                    // been concurrently uploaded by another client. In this case, they return a
-                    // WriteResponse with a committed_size equal to the digest's entire size before
-                    // closing the stream.
-                    // Because the server then closes the stream, the client gets an RpcFinished
-                    // error in this case. We ignore this, and will later on verify that the
-                    // committed_size we received from the server is equal to the expected one. If
-                    // these are not equal, the upload will be considered a failure at that point.
-                    // Whether this type of response will become part of the official API is up for
-                    // discussion: see
-                    // https://groups.google.com/d/topic/remote-execution-apis/NXUe3ItCw68/discussion.
-                    grpcio::Error::RpcFinished(None) => Ok(()),
-                    e => Err(format!(
-                      "Error attempting to upload digest {:?}: {:?}",
-                      digest, e
-                    )),
-                  }
-                })
-                .and_then(move |()| {
-                  receiver.map_err(move |e| {
-                    format!(
-                      "Error from server when uploading digest {:?}: {:?}",
-                      digest, e
-                    )
-                  })
-                })
-                .and_then(move |received| {
-                  if received.get_committed_size() == len as i64 {
-                    Ok(digest)
-                  } else {
-                    Err(format!(
-                      "Uploading file with digest {:?}: want commited size {} but got {}",
-                      digest,
-                      len,
-                      received.get_committed_size()
-                    ))
-                  }
-                })
-                .to_boxed()
-            }
-          }
-        })
-        .then(move |future| {
-          let workunit = workunit_store::WorkUnit {
-            name: workunit_name.clone(),
-            time_span: TimeSpan::since(&start_time),
-            span_id: workunit_store::generate_random_64bit_string(),
-            parent_id: workunit_store::get_parent_id(),
-          };
-          workunit_store.add_workunit(workunit);
-          future
-        })
-        .to_boxed()
-    }
-
-    pub fn load_bytes_with<T: Send + 'static, F: Fn(Bytes) -> T + Send + Sync + Clone + 'static>(
-      &self,
-      _entry_type: EntryType,
-      digest: Digest,
-      f: F,
-      workunit_store: WorkUnitStore,
-    ) -> BoxFuture<Option<T>, String> {
-      let start_time = std::time::SystemTime::now();
-
-      let store = self.clone();
-      let resource_name = format!(
-        "{}/blobs/{}/{}",
-        store.instance_name.clone().unwrap_or_default(),
-        digest.0,
-        digest.1
-      );
-      let workunit_name = format!("load_bytes_with({})", resource_name.clone());
-      let workunit_store = workunit_store.clone();
-      self
-        .with_byte_stream_client(move |client| {
-          match client
-            .read_opt(
-              &{
-                let mut req = bazel_protos::bytestream::ReadRequest::new();
-                req.set_resource_name(resource_name.clone());
-                req.set_read_offset(0);
-                // 0 means no limit.
-                req.set_read_limit(0);
-                req
-              },
-              store.call_option(),
-            )
-            .map(|stream| (stream, client))
-          {
-            Ok((stream, client)) => {
-              let f = f.clone();
-              // We shouldn't have to pass around the client here, it's a workaround for
-              // https://github.com/pingcap/grpc-rs/issues/123
-              future::ok(client)
-                .join(
-                  stream.fold(BytesMut::with_capacity(digest.1), move |mut bytes, r| {
-                    bytes.extend_from_slice(&r.data);
-                    future::ok::<_, grpcio::Error>(bytes)
-                  }),
-                )
-                .map(|(_client, bytes)| Some(bytes.freeze()))
-                .or_else(|e| match e {
-                  grpcio::Error::RpcFailure(grpcio::RpcStatus {
-                    status: grpcio::RpcStatusCode::NotFound,
-                    ..
-                  }) => Ok(None),
-                  _ => Err(format!(
-                    "Error from server in response to CAS read request: {:?}",
-                    e
-                  )),
-                })
-                .map(move |maybe_bytes| maybe_bytes.map(f))
-                .to_boxed()
-            }
-            Err(err) => future::err(format!(
-              "Error making CAS read request for {:?}: {:?}",
-              digest, err
-            ))
-            .to_boxed(),
-          }
-        })
-        .then(move |future| {
-          let workunit = workunit_store::WorkUnit {
-            name: workunit_name.clone(),
-            time_span: TimeSpan::since(&start_time),
-            span_id: workunit_store::generate_random_64bit_string(),
-            parent_id: workunit_store::get_parent_id(),
-          };
-          workunit_store.add_workunit(workunit);
-          future
-        })
-        .to_boxed()
-    }
-
-    ///
-    /// Given a collection of Digests (digests),
-    /// returns the set of digests from that collection not present in the CAS.
-    ///
-    pub fn list_missing_digests(
-      &self,
-      request: bazel_protos::remote_execution::FindMissingBlobsRequest,
-      workunit_store: WorkUnitStore,
-    ) -> impl Future<Item = HashSet<Digest>, Error = String> {
-      let start_time = std::time::SystemTime::now();
-
-      let store = self.clone();
-      let workunit_name = format!(
-        "list_missing_digests({})",
-        store.instance_name.clone().unwrap_or_default()
-      );
-      let workunit_store = workunit_store.clone();
-      self
-        .with_cas_client(move |client| {
-          client
-            .find_missing_blobs_opt(&request, store.call_option())
-            .map_err(|err| {
-              format!(
-                "Error from server in response to find_missing_blobs_request: {:?}",
-                err
-              )
-            })
-            .and_then(|response| {
-              response
-                .get_missing_blob_digests()
-                .iter()
-                .map(|digest| digest.into())
-                .collect()
-            })
-        })
-        .then(move |future| {
-          let workunit = workunit_store::WorkUnit {
-            name: workunit_name.clone(),
-            time_span: TimeSpan::since(&start_time),
-            span_id: workunit_store::generate_random_64bit_string(),
-            parent_id: workunit_store::get_parent_id(),
-          };
-          workunit_store.add_workunit(workunit);
-          future
-        })
-    }
-
-    pub(super) fn find_missing_blobs_request<'a, Digests: Iterator<Item = &'a Digest>>(
-      &self,
-      digests: Digests,
-    ) -> bazel_protos::remote_execution::FindMissingBlobsRequest {
-      let mut request = bazel_protos::remote_execution::FindMissingBlobsRequest::new();
-      if let Some(ref instance_name) = self.instance_name {
-        request.set_instance_name(instance_name.clone());
-      }
-      for digest in digests {
-        request.mut_blob_digests().push(digest.into());
-      }
-      request
-    }
-  }
-
-  #[cfg(test)]
-  mod tests {
-    use super::super::{EntryType, MEGABYTES};
-    use super::ByteStore;
-    use bytes::Bytes;
-    use hashing::Digest;
-    use mock::StubCAS;
-    use serverset::BackoffConfig;
-    use std::collections::HashSet;
-    use std::time::Duration;
-    use testutil::data::{TestData, TestDirectory};
-    use workunit_store::WorkUnitStore;
-
-    use super::super::tests::{
-      big_file_bytes, big_file_digest, big_file_fingerprint, block_on, new_cas,
-    };
-
-    #[test]
-    fn loads_file() {
-      let testdata = TestData::roland();
-      let cas = new_cas(10);
-
-      assert_eq!(
-        load_file_bytes(&new_byte_store(&cas), testdata.digest()).unwrap(),
-        Some(testdata.bytes())
-      );
-    }
-
-    #[test]
-    fn missing_file() {
-      let cas = StubCAS::empty();
-
-      assert_eq!(
-        load_file_bytes(&new_byte_store(&cas), TestData::roland().digest()),
-        Ok(None)
-      );
-    }
-
-    #[test]
-    fn load_directory() {
-      let cas = new_cas(10);
-      let testdir = TestDirectory::containing_roland();
-
-      assert_eq!(
-        load_directory_proto_bytes(&new_byte_store(&cas), testdir.digest()),
-        Ok(Some(testdir.bytes()))
-      );
-    }
-
-    #[test]
-    fn missing_directory() {
-      let cas = StubCAS::empty();
-
-      assert_eq!(
-        load_directory_proto_bytes(
-          &new_byte_store(&cas),
-          TestDirectory::containing_roland().digest()
-        ),
-        Ok(None)
-      );
-    }
-
-    #[test]
-    fn load_file_grpc_error() {
-      let cas = StubCAS::always_errors();
-
-      let error = load_file_bytes(&new_byte_store(&cas), TestData::roland().digest())
-        .expect_err("Want error");
-      assert!(
-        error.contains("StubCAS is configured to always fail"),
-        format!("Bad error message, got: {}", error)
-      )
-    }
-
-    #[test]
-    fn load_directory_grpc_error() {
-      let cas = StubCAS::always_errors();
-
-      let error = load_directory_proto_bytes(
-        &new_byte_store(&cas),
-        TestDirectory::containing_roland().digest(),
-      )
-      .expect_err("Want error");
-      assert!(
-        error.contains("StubCAS is configured to always fail"),
-        format!("Bad error message, got: {}", error)
-      )
-    }
-
-    #[test]
-    fn fetch_less_than_one_chunk() {
-      let testdata = TestData::roland();
-      let cas = new_cas(testdata.bytes().len() + 1);
-
-      assert_eq!(
-        load_file_bytes(&new_byte_store(&cas), testdata.digest()),
-        Ok(Some(testdata.bytes()))
-      )
-    }
-
-    #[test]
-    fn fetch_exactly_one_chunk() {
-      let testdata = TestData::roland();
-      let cas = new_cas(testdata.bytes().len());
-
-      assert_eq!(
-        load_file_bytes(&new_byte_store(&cas), testdata.digest()),
-        Ok(Some(testdata.bytes()))
-      )
-    }
-
-    #[test]
-    fn fetch_multiple_chunks_exact() {
-      let testdata = TestData::roland();
-      let cas = new_cas(1);
-
-      assert_eq!(
-        load_file_bytes(&new_byte_store(&cas), testdata.digest()),
-        Ok(Some(testdata.bytes()))
-      )
-    }
-
-    #[test]
-    fn fetch_multiple_chunks_nonfactor() {
-      let testdata = TestData::roland();
-      let cas = new_cas(9);
-
-      assert_eq!(
-        load_file_bytes(&new_byte_store(&cas), testdata.digest()),
-        Ok(Some(testdata.bytes()))
-      )
-    }
-
-    #[test]
-    fn write_file_one_chunk() {
-      let testdata = TestData::roland();
-      let cas = StubCAS::empty();
-
-      let store = new_byte_store(&cas);
-      assert_eq!(
-        block_on(store.store_bytes(testdata.bytes(), WorkUnitStore::new())),
-        Ok(testdata.digest())
-      );
-
-      let blobs = cas.blobs.lock();
-      assert_eq!(blobs.get(&testdata.fingerprint()), Some(&testdata.bytes()));
-    }
-
-    #[test]
-    fn write_file_multiple_chunks() {
-      let cas = StubCAS::empty();
-
-      let store = ByteStore::new(
-        vec![cas.address()],
-        None,
-        None,
-        None,
-        1,
-        10 * 1024,
-        Duration::from_secs(5),
-        BackoffConfig::new(Duration::from_millis(10), 1.0, Duration::from_millis(10)).unwrap(),
-        1,
-        1,
-      )
-      .unwrap();
-
-      let all_the_henries = big_file_bytes();
-
-      let fingerprint = big_file_fingerprint();
-
-      assert_eq!(
-        block_on(store.store_bytes(all_the_henries.clone(), WorkUnitStore::new())),
-        Ok(big_file_digest())
-      );
-
-      let blobs = cas.blobs.lock();
-      assert_eq!(blobs.get(&fingerprint), Some(&all_the_henries));
-
-      let write_message_sizes = cas.write_message_sizes.lock();
-      assert_eq!(
-        write_message_sizes.len(),
-        98,
-        "Wrong number of chunks uploaded"
-      );
-      for size in write_message_sizes.iter() {
-        assert!(
-          size <= &(10 * 1024),
-          format!("Size {} should have been <= {}", size, 10 * 1024)
-        );
-      }
-    }
-
-    #[test]
-    fn write_empty_file() {
-      let empty_file = TestData::empty();
-      let cas = StubCAS::empty();
-
-      let store = new_byte_store(&cas);
-      assert_eq!(
-        block_on(store.store_bytes(empty_file.bytes(), WorkUnitStore::new())),
-        Ok(empty_file.digest())
-      );
-
-      let blobs = cas.blobs.lock();
-      assert_eq!(
-        blobs.get(&empty_file.fingerprint()),
-        Some(&empty_file.bytes())
-      );
-    }
-
-    #[test]
-    fn write_file_errors() {
-      let cas = StubCAS::always_errors();
-
-      let store = new_byte_store(&cas);
-      let error = block_on(store.store_bytes(TestData::roland().bytes(), WorkUnitStore::new()))
-        .expect_err("Want error");
-      assert!(
-        error.contains("Error from server"),
-        format!("Bad error message, got: {}", error)
-      );
-      assert!(
-        error.contains("StubCAS is configured to always fail"),
-        format!("Bad error message, got: {}", error)
-      );
-    }
-
-    #[test]
-    fn write_connection_error() {
-      let store = ByteStore::new(
-        vec![String::from("doesnotexist.example")],
-        None,
-        None,
-        None,
-        1,
-        10 * 1024 * 1024,
-        Duration::from_secs(1),
-        BackoffConfig::new(Duration::from_millis(10), 1.0, Duration::from_millis(10)).unwrap(),
-        1,
-        1,
-      )
-      .unwrap();
-      let error = block_on(store.store_bytes(TestData::roland().bytes(), WorkUnitStore::new()))
-        .expect_err("Want error");
-      assert!(
-        error.contains("Error attempting to upload digest"),
-        format!("Bad error message, got: {}", error)
-      );
-    }
-
-    #[test]
-    fn list_missing_digests_none_missing() {
-      let cas = new_cas(1024);
-
-      let store = new_byte_store(&cas);
-      assert_eq!(
-        block_on(store.list_missing_digests(
-          store.find_missing_blobs_request(vec![TestData::roland().digest()].iter()),
-          WorkUnitStore::new(),
-        )),
-        Ok(HashSet::new())
-      );
-    }
-
-    #[test]
-    fn list_missing_digests_some_missing() {
-      let cas = StubCAS::empty();
-
-      let store = new_byte_store(&cas);
-
-      let digest = TestData::roland().digest();
-
-      let mut digest_set = HashSet::new();
-      digest_set.insert(digest);
-
-      assert_eq!(
-        block_on(store.list_missing_digests(
-          store.find_missing_blobs_request(vec![digest].iter()),
-          WorkUnitStore::new(),
-        )),
-        Ok(digest_set)
-      );
-    }
-
-    #[test]
-    fn list_missing_digests_error() {
-      let cas = StubCAS::always_errors();
-
-      let store = new_byte_store(&cas);
-
-      let error = block_on(store.list_missing_digests(
-        store.find_missing_blobs_request(vec![TestData::roland().digest()].iter()),
-        WorkUnitStore::new(),
-      ))
-      .expect_err("Want error");
-      assert!(
-        error.contains("StubCAS is configured to always fail"),
-        format!("Bad error message, got: {}", error)
-      );
-    }
-
-    #[test]
-    fn reads_from_multiple_cas_servers() {
-      let roland = TestData::roland();
-      let catnip = TestData::catnip();
-
-      let cas1 = StubCAS::builder().file(&roland).file(&catnip).build();
-      let cas2 = StubCAS::builder().file(&roland).file(&catnip).build();
-
-      let store = ByteStore::new(
-        vec![cas1.address(), cas2.address()],
-        None,
-        None,
-        None,
-        1,
-        10 * 1024 * 1024,
-        Duration::from_secs(1),
-        BackoffConfig::new(Duration::from_millis(10), 1.0, Duration::from_millis(10)).unwrap(),
-        1,
-        2,
-      )
-      .unwrap();
-
-      assert_eq!(
-        load_file_bytes(&store, roland.digest()),
-        Ok(Some(roland.bytes()))
-      );
-
-      assert_eq!(
-        load_file_bytes(&store, catnip.digest()),
-        Ok(Some(catnip.bytes()))
-      );
-
-      assert_eq!(cas1.read_request_count(), 1);
-      assert_eq!(cas2.read_request_count(), 1);
-    }
-
-    fn new_byte_store(cas: &StubCAS) -> ByteStore {
-      ByteStore::new(
-        vec![cas.address()],
-        None,
-        None,
-        None,
-        1,
-        10 * MEGABYTES,
-        Duration::from_secs(1),
-        BackoffConfig::new(Duration::from_millis(10), 1.0, Duration::from_millis(10)).unwrap(),
-        1,
-        1,
-      )
-      .unwrap()
-    }
-
-    pub fn load_file_bytes(store: &ByteStore, digest: Digest) -> Result<Option<Bytes>, String> {
-      load_bytes(&store, EntryType::File, digest)
-    }
-
-    pub fn load_directory_proto_bytes(
-      store: &ByteStore,
-      digest: Digest,
-    ) -> Result<Option<Bytes>, String> {
-      load_bytes(&store, EntryType::Directory, digest)
-    }
-
-    fn load_bytes(
-      store: &ByteStore,
-      entry_type: EntryType,
-      digest: Digest,
-    ) -> Result<Option<Bytes>, String> {
-      block_on(store.load_bytes_with(entry_type, digest, |b| b, WorkUnitStore::new()))
-    }
-  }
 }
 
 #[cfg(test)]
